Replace old StrategicChargingReservationEngine with DSim ready version.
